[WIP][JIT] Add JIT backend registration API
Summary:
This commit adds `torch::jit::RegisterBackend`, an API that allows
external backends to be registered for the execution of JIT subgraphs
outside the JIT interpreter. In order to register an external backend,
functions for preprocessing, compiling, and executing a Module must be
provided. Then, a `ScriptModule` that can compile and execute a given JIT
subgraph using the functions provided at registration time is generated
for each registered backend.

Testing:
This commit adds a unit test that uses a test backend that does little to
no actual work to make sure that the registration endpoints and generated
`ScriptModule` actually work.